test(DB-17): 작성한 리뷰 전체 보기 API : Controller Unit Test 작성

+++ dongsan-api/src/main/java/com/dongsan/domains/user/usecase/UserReviewUseCase.java
@@ -15,7 +15,7 @@ public class UserReviewUseCase {
     private final ReviewQueryService reviewQueryService;
 
     @Transactional(readOnly = true)
-    public GetReview getReviews(Integer limit, Integer reviewId, Long memberId) {
+    public GetReview getReviews(Integer limit, Long reviewId, Long memberId) {
         List<Review> reviews = reviewQueryService.getReviews(limit, reviewId, memberId);
         return UserReviewMapper.toGetReview(reviews);
     }
