+++ dongsan-domain/src/main/java/com/dongsan/domains/review/repository/ReviewQueryDSLRepository.java
@@ -16,7 +16,7 @@ public class ReviewQueryDSLRepository{
     private QReview review = QReview.review;
 
     // TODO : 쿼리 확인 필요 (member join 안해도 쿼리 발생 하는지, 안하는지)
-    public List<Review> getReviews(Integer limit, Integer reviewId, Long memberId) {
+    public List<Review> getReviews(Integer limit, Long reviewId, Long memberId) {
         return queryFactory.selectFrom(review)
                 //.join(review.member)
                 .join(review.walkway).fetchJoin()
@@ -26,7 +26,7 @@ public class ReviewQueryDSLRepository{
                 .fetch();
     }
 
-    private BooleanExpression reviewIdLt(Integer reviewId){
+    private BooleanExpression reviewIdLt(Long reviewId){
         // 조건 만족 안하면 null 반환
         // where 절에서 null은 무시된다.
         return reviewId != null ? review.id.lt(reviewId) : null;
